+++ CHANGELOG.md
@@ -25,6 +25,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `RawJson` format which enables to retrieve events in raw format while also getting the metadata added by OpenWEC
 - Add the subscription revision in OpenWEC events metadata
 - Add `locale` and `data_locale` subscriptions parameters
+- Add support for Proxy Protocol to allow openwec to be used behind a layer 4 load
+balancer whilst preserving the client IP address and port.
 
 ### Changed